[Core] Move postgres to /postgres folder
