+++ routes/contract.js
@@ -1,11 +1,11 @@
 var express = require('express');
-var contract = require('../models/contract.js');
-var contractFunction = require('../models/contractFunction.js');
-var contractEvent = require('../models/contractEvent.js');
+var contract = require('../models/postgres/contract.js');
+var contractFunction = require('../models/postgres/contractFunction.js');
+var contractEvent = require('../models/postgres/contractEvent.js');
 var contractController = require('../controllers/contractController.js');
-var transactionData = require('../models/transactionData.js');
-var eventData = require('../models/eventData.js');
-var webhookData = require('../models/webhookData.js');
+var transactionData = require('../models/postgres/transactionData.js');
+var eventData = require('../models/postgres/eventData.js');
+var webhookData = require('../models/postgres/webhookData.js');
 var sqsHelper = require('../helpers/aws/sqsHelper.js');
 var request = require('request');
 var router = express.Router();
