+++ routes/ethereum.js
@@ -3,10 +3,10 @@ var Web3 = require('web3');
 var solc = require('solc');
 var util = require('util');
 var Promise = require('bluebird');
-var contract = require('../models/contract.js');
-var contractFunction = require('../models/contractFunction.js');
-var contractEvent = require('../models/contractEvent.js');
-var account = require('../models/account.js');
+var contract = require('../models/postgres/contract.js');
+var contractFunction = require('../models/postgres/contractFunction.js');
+var contractEvent = require('../models/postgres/contractEvent.js');
+var account = require('../models/postgres/account.js');
 var ethereumController = require('../controllers/ethereumController.js');
 var contractController = require('../controllers/contractController.js');
 var sqsHelper = require('../helpers/aws/sqsHelper.js');
