+++ schedulers/aws/sqsEventConsumer.js
@@ -1,9 +1,9 @@
 var Consumer = require('sqs-consumer');
 var AWS = require('aws-sdk');
 var Web3 = require('web3');
-var contract = require('../../models/contract.js');
-var contractEvent = require('../../models/contractEvent.js');
-var eventData = require('../../models/eventData.js');
+var contract = require('../../models/postgres/contract.js');
+var contractEvent = require('../../models/postgres/contractEvent.js');
+var eventData = require('../../models/postgres/eventData.js');
 var sqsHelper = require('../../helpers/aws/sqsHelper.js');
 var web3 = new Web3();
