+++ schedulers/aws/sqsTransactionReceiptConsumer.js
@@ -1,10 +1,10 @@
 var Consumer = require('sqs-consumer');
 var AWS = require('aws-sdk');
 var Web3 = require('web3');
-var contract = require('../../models/contract.js');
-var contractFunction = require('../../models/contractFunction.js');
-var contractEvent = require('../../models/contractEvent.js');
-var transactionData = require('../../models/transactionData.js');
+var contract = require('../../models/postgres/contract.js');
+var contractFunction = require('../../models/postgres/contractFunction.js');
+var contractEvent = require('../../models/postgres/contractEvent.js');
+var transactionData = require('../../models/postgres/transactionData.js');
 var EventListenerHelper = require('../../helpers/eventListenerHelper.js');
 var sqsHelper = require('../../helpers/aws/sqsHelper.js');
 var web3 = new Web3();
