+++ tests/others/solcTest.js
@@ -2,9 +2,9 @@ var solc = require('solc');
 var util = require('util');
 var Web3 = require('web3');
 var sqsHelper = require('../../helpers/aws/sqsHelper.js');
-var contract = require('../../models/contract.js');
-var contractFunction = require('../../models/contractFunction.js');
-var contractEvent = require('../../models/contractEvent.js');
+var contract = require('../../models/postgres/contract.js');
+var contractFunction = require('../../models/postgres/contractFunction.js');
+var contractEvent = require('../../models/postgres/contractEvent.js');
 var web3 = new Web3();
 
 web3.setProvider(new web3.providers.HttpProvider(process.env.ENODE_BASE || 'http://localhost:8545'));
